Do not run FAB pre-commits in release branches
The openapi generation scripts were run together for FAB and main
openapi from core - however we do not want to re-run those scripts at
all in non-main branch - i.e. branch we prepare providers from.

This PR splits the openapi script and removes all fab pre-commits from
running in non-main branch.
