+++ dev/breeze/doc/images/output_static-checks.txt
@@ -1,1 +1,1 @@
-d9df5f04cf2495051df6c928cf70d111
+44ddf285d8b0b2f0369d4750ace7dde9
