+++ scripts/ci/pre_commit/generate_openapi_spec.py
@@ -30,7 +30,7 @@ from common_precommit_utils import (
 initialize_breeze_precommit(__name__, __file__)
 
 cmd_result = run_command_via_breeze_shell(
-    ["python3", "/opt/airflow/scripts/in_container/run_update_fastapi_api_spec.py"],
+    ["python3", "/opt/airflow/scripts/in_container/run_generate_openapi_spec.py"],
     backend="postgres",
     skip_environment_initialization=False,
 )
