+++ scripts/in_container/in_container_utils.py
@@ -18,12 +18,18 @@ from __future__ import annotations
 
 import shlex
 import subprocess
+import sys
 import textwrap
 from contextlib import contextmanager
 from pathlib import Path
+from typing import TYPE_CHECKING
 
 import rich_click as click
 from rich.console import Console
+
+if TYPE_CHECKING:
+    from fastapi import FastAPI
+
 
 click.rich_click.COLOR_SYSTEM = "standard"
 console = Console(width=400, color_system="standard")
@@ -56,3 +62,48 @@ def run_command(cmd: list[str], github_actions: bool, **kwargs) -> subprocess.Co
         console.print(f"[red]Command failed: {' '.join([shlex.quote(entry) for entry in cmd])}[/]")
         console.print("[red]Please unfold the above group and to investigate the issue[/]")
     return result
+
+
+def generate_openapi_file(app: FastAPI, file_path: Path, prefix: str = "", only_ui: bool = False):
+    import yaml
+    from fastapi.openapi.utils import get_openapi
+    from fastapi.routing import APIRoute
+
+    if only_ui:
+        for route in app.routes:
+            if not isinstance(route, APIRoute):
+                continue
+            route.include_in_schema = route.path.startswith("/ui/")
+
+    with file_path.open("w+") as f:
+        openapi_schema = get_openapi(
+            title=app.title,
+            version=app.version,
+            openapi_version=app.openapi_version,
+            description=app.description,
+            routes=app.routes,
+        )
+        if prefix:
+            openapi_schema["paths"] = {
+                prefix + path: path_dict for path, path_dict in openapi_schema["paths"].items()
+            }
+        yaml.dump(
+            openapi_schema,
+            f,
+            default_flow_style=False,
+            sort_keys=False,
+        )
+
+
+def validate_openapi_file(file_path: Path) -> bool:
+    import yaml
+    from openapi_spec_validator import validate_spec
+
+    with file_path.open() as f:
+        openapi_schema = yaml.safe_load(f)
+    try:
+        validate_spec(openapi_schema)
+    except Exception as e:
+        print(f"[ERROR] OpenAPI validation failed for {file_path}: {e}", file=sys.stderr)
+        sys.exit(1)
+    return True
